Fold the two copies of uploadAttachments into one
The web and mobile copies had drifted only in how they dropped a failed
upload, so shared gets the web one and both clients import it.
